(tests) .travis.yml: don't use Docker
Tests are too fast for the start time of "sudo: required" to be appropriate.

If we needed to run the tests really often, it would be necessary not to impose
load on third-party server, but this project is not in active development,
so the number of requests would be very small.